delete mainviewmodel moved all to ResultViewModel2 refactored bottompreview
