+++ app/src/main/java/com/lagradost/quicknovel/ui/result/ResultFragment2.kt
@@ -4,21 +4,23 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.lagradost.quicknovel.R
 import com.lagradost.quicknovel.compose.CloudStreamTheme
-import com.lagradost.quicknovel.compose.ObserveEffect
 import com.lagradost.quicknovel.compose.loadPrimaryColor
 import com.lagradost.quicknovel.compose.loadThemeMode
 import com.lagradost.quicknovel.util.UIHelper.colorFromAttribute
 
 class ResultFragment2 : Fragment() {
+    private val viewModel: ResultViewModel2 by activityViewModels()
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
@@ -27,19 +29,28 @@ class ResultFragment2 : Fragment() {
         setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
 
         setContent {
-            val viewModel: ResultViewModel2 =
-                viewModel(factory = ResultViewModel2.provideFactory(requireArguments()))
-
             CloudStreamTheme(
                 mode = LocalContext.current.loadThemeMode(),
                 primaryColor = LocalContext.current.loadPrimaryColor(),
             ) {
                 val state by viewModel.state.collectAsStateWithLifecycle()
 
-                ObserveEffect(viewModel.effect) { _ ->
-                    // Not yet implemented
+                LaunchedEffect(Unit) {
+                    val url = arguments?.getString("url")
+                    val apiName = arguments?.getString("apiName")
+                    val id = arguments?.getInt("id", -1)?.takeIf { it != -1 }
+
+                    if (url != null && apiName != null) {
+                        viewModel.onAction(
+                            ResultPageAction.LoadResult(
+                                url = url,
+                                apiName = apiName,
+                                id = id
+                            )
+                        )
+                    }
                 }
-                ResultScreen(state,viewModel::onAction)
+                ResultScreen(state, viewModel::onAction)
             }
         }
     }
